Reserve negative id's for DDMS.
On android DDMS traffic is multiplexed over the same connection that
jdwp uses and transmits using the same format. In order to avoid
confusing ddms clients we need to ensure that DDMS packets do not use
ids that conflict with jdwp packets.

In order to prevent these collisions we reserve all negative id
numbers for DDMS use and prevent the libjdwp agent from sending
packets with these id numbers.

The reduction in truly unique id numbers from 4294967296 to 2147483648
is not expected to be a problem.

Test: ./art/tools/run-libjdwp-tests.sh --mode=host
Bug: 62821960
Change-Id: Ia05ae056ed7fc6b7b87b0a1d7785eb0653e8356f