Prevent class-load & class prepare events from being sent on arrays
Due to b/111394423 ART will sometimes send array-class (and only
array-class) ClassLoaded and ClassPrepared events on JIT worker
threads. It is possible for the process to deadlock if libjdwp
attempts to gain any locks while handling these events in jit-threads.
Since array-class loads and prepares are not meant to be sent for
these events anyway, we can simply bail out early if we see an array
class in these events. This both fixes a (small) incompatibility
between this agent running on the RI and on ART and prevents the
deadlocks running code on the jit thread can cause.

Bug: 111394423
Test: ./art/tools/run-libjdwp-tests.sh --mode=host
Change-Id: Ied7851c25862f61cf812d384f57a609b1bed30f7